project is now building and pass ci tests

+++ truck/src/main/java/com/polytech/si5/al/dronedelivery/team/g/truck/utils/PositionCalculator.java
@@ -5,10 +5,10 @@ import com.polytech.si5.al.dronedelivery.team.g.truck.entities.Position;
 public class PositionCalculator {
 
     public static double distance(Position p1, Position p2){
-        int xP1 = p1.getLongitude();
-        int yP1 = p1.getLatitude();
-        int xP2 = p2.getLongitude();
-        int yP2 = p2.getLatitude();
+        double xP1 = p1.getLongitude();
+        double yP1 = p1.getLatitude();
+        double xP2 = p2.getLongitude();
+        double yP2 = p2.getLatitude();
 
         return Math.abs(xP1 - xP2) + Math.abs(yP1 - yP2);
     }
